chore: move spoof module to ethers_core
