+++ examples/transactions/examples/trace_call.rs
@@ -26,6 +26,7 @@ async fn main() -> Result<()> {
                 )),
                 ..Default::default()
             },
+            state_overrides: None,
         };
         let traces = client.debug_trace_call(tx, Some(block), options).await?;
         println!("{traces:?}");
